requirements.txt
flask 
flask-login 
flask-sqlalchemy 
flask-wtf
python-dotenv